feat: progress towards metadata rewriting
Metadata maps now use an arena allocator to deepcopy the keys and
values, which avoids any use-after-free from truncating files.

Turns out there are a number of weird exceptions when it comes to
writing metadata. There can be additional `/Past OFFSET` entries in
trailers which I have no means of dealing with at the moment that point
to different xref tables. There's also just weird quirks related to
compression that I'm not sure how to deal with.