feat: new metadata map value type
The metadata map can now be used to track the post-script type used in
the PDF file. This allows writing the correct type into the output (e.g.
strings, list, literals).

Used to update the map after the trailer to hold the correct `/Prev`
offset value for the cross-reference table.